Split out Glsl/GlslStruct traits, stub out custom derive
