+++ crevice-derive/src/glsl.rs
@@ -1,0 +1,16 @@
+use proc_macro2::{Span, TokenStream};
+use quote::{format_ident, quote};
+use syn::{parse_quote, Data, DeriveInput, Fields, Ident, Path};
+
+pub fn emit(input: DeriveInput) -> TokenStream {
+    let fields = match &input.data {
+        Data::Struct(data) => match &data.fields {
+            Fields::Named(fields) => fields,
+            Fields::Unnamed(_) => panic!("Tuple structs are not supported"),
+            Fields::Unit => panic!("Unit structs are not supported"),
+        },
+        Data::Enum(_) | Data::Union(_) => panic!("Only structs are supported"),
+    };
+
+    todo!()
+}
